// src/types.ts
export type SertifikatTanahType = 'SHM' | 'WAQAF' | 'SHGB' | 'LETER C';

export interface ProgresBalikNama {
  id: string;
  tanggal: string;
  keterangan: string;
  biaya: number;
}

export interface RiwayatBalikNama {
  tanggalSelesai: string;
  atasNamaLama: string;
  nomerSertifikatLama: string;
  catatanLama?: string;
  nomorSertifikatBaru: string;
  progresDetail?: ProgresBalikNama[];
}

export interface PeminjamanBerkas {
  id?: string;
  peminjamName: string;
  peminjamJabatan: string;
  peminjamKontak: string;
  tanggalPinjam: string;
  keperluan: string;
  tanggalKembaliRencana?: string;
  tanggalKembaliRiil?: string;
  status: 'DIPINJAM' | 'DIKEMBALIKAN';
  namaPetugas: string;
  fotoPinjam?: string;
  fotoKembali?: string;
}

export interface AsetTanah {
  id: string;
  type: 'tanah';
  jenisSertifikat: SertifikatTanahType;
  nomerSertifikat: string;
  atasNamaSertifikat: string;
  lokasi: string;
  penggunaan: string;
  tempatSimpanBerkas: string;
  koordinatGoogleMaps?: string;
  luasTanah?: number; // luas tanah dlm m2
  createdAt: number;
  sedangBalikNama?: boolean;
  namaPemilikBaru?: string;
  catatanBalikNama?: string;
  tanggalMulaiBalikNama?: string;
  progresBalikNama?: ProgresBalikNama[];
  riwayatBalikNama?: RiwayatBalikNama[];
  sedangDipinjam?: boolean;
  peminjamanAktif?: PeminjamanBerkas;
  riwayatPeminjaman?: PeminjamanBerkas[];
}

export type JenisKendaraanType = 'MOTOR' | 'MOBIL' | 'ELF' | 'BUS';

export type KondisiType = 'BAIK' | 'RUSAK RINGAN' | 'RUSAK BERAT';

export interface AsetKendaraan {
  id: string;
  type: 'kendaraan';
  jenisKendaraan: JenisKendaraanType;
  nomorPolisi: string;
  merk: string;
  atasNama: string;
  tahunPembuatan: number;
  kondisiKendaraan: KondisiType;
  tanggalBulanPajak?: string; // Tanggal & Bulan Pajak
  penanggungJawabDaerah?: string; // Penanggung Jawab Daerah Dropdown (Desa / Kelompok)
  createdAt: number;
  sedangBalikNama?: boolean;
  namaPemilikBaru?: string;
  catatanBalikNama?: string;
  tanggalMulaiBalikNama?: string;
  progresBalikNama?: ProgresBalikNama[];
  riwayatBalikNama?: RiwayatBalikNama[];
  sedangDipinjam?: boolean;
  peminjamanAktif?: PeminjamanBerkas;
  riwayatPeminjaman?: PeminjamanBerkas[];
}
// ...
export interface AsetBangunan {
  id: string;
  type: 'bangunan';
  namaBangunan: string;
  lokasi: string;
  luasBangunan: number; // in m2
  koordinatGoogleMaps?: string;
  penggunaanBangunan: string;
  nomerPBG: string;
  nomerSLF: string;
  kondisi: KondisiType;
  keteranganKerusakan: string;
  createdAt: number;
  sedangBalikNama?: boolean;
  namaPemilikBaru?: string;
  catatanBalikNama?: string;
  tanggalMulaiBalikNama?: string;
  progresBalikNama?: ProgresBalikNama[];
  riwayatBalikNama?: RiwayatBalikNama[];
  sedangDipinjam?: boolean;
  peminjamanAktif?: PeminjamanBerkas;
  riwayatPeminjaman?: PeminjamanBerkas[];
}

export type Asset = AsetTanah | AsetKendaraan | AsetBangunan;